requirements.txt
fastapi>=0.114.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.25
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.0
python-multipart>=0.0.9
jinja2>=3.1.2
email-validator>=2.0.0
google-generativeai>=0.3.0
apscheduler>=3.10.4
pytz>=2023.3
sendgrid>=6.10.0
python-dotenv>=1.0.0